+++ R/export2word.R
@@ -1,5 +1,7 @@
-export2word <- function(x, file, which.table="descr", nmax = TRUE, header.labels = c()){
-
+export2word <- function(x, file, which.table="descr", nmax=TRUE, header.labels=c(), caption=NULL, 
+                        strip=FALSE, first.strip=FALSE, background="#D2D2D2", size=NULL,  
+                        header.background=NULL, header.color=NULL){
+  
   if (!inherits(x, "createTable")) 
     stop("x must be of class 'createTable'")
   # if (inherits(x, "cbind.createTable")) 
@@ -20,7 +22,10 @@ export2word <- function(x, file, which.table="descr", nmax = TRUE, header.labels
 \n\n\n
 
 ```{r, echo=FALSE}\n
-export2md(x, which.table, nmax , header.labels, format='markdown')\n
+export2md(x, which.table=which.table, nmax=nmax, header.labels=header.labels, 
+caption=caption, format='markdown', strip=strip, 
+first.strip=first.strip, background=background, size=size, 
+header.background=header.background, header.color=header.color)\n
 ```\n
 \n"
 ,sep=""